Renaming github-jenkins to github
