+++ schema/openapi.json
@@ -470,17 +470,17 @@
         }
       }
     },
-    "/github-jenkins": {
+    "/github": {
       "post": {
         "description": "Create a new Github commit from Jenkins job",
-        "operationId": "post.githubJenkins",
+        "operationId": "post.github",
         "requestBody": {
           "description": "Simple JSON Data from commit Jenkins job",
           "required": true,
           "content": {
             "appplication/json": {
               "schema": {
-                "$ref": "#/components/schemas/JenkinsPayload"
+                "$ref": "#/components/schemas/GithubPayload"
               }
             }
           }
@@ -699,7 +699,7 @@
           "msg"
         ]
       },
-      "JenkinsPayload": {
+      "GithubPayload": {
         "type": "object",
         "properties": {
           "app": {
